requirements.txt
openai>=1.30.1
moviepy>=1.0.3
Pillow>=10.3.0
edge-tts>=6.1.1
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.135.0

/* "voice": "en-US-GuyNeural", */